Refactors UI code of integrated circuits completely into separate folders and adds dragndrop functionality for the UI
