+++ tgui/packages/tgui/interfaces/IntegratedCircuit/constants.js
@@ -1,0 +1,5 @@
+export const NULL_REF = '[0x0]';
+export const ABSOLUTE_Y_OFFSET = -32;
+export const SVG_CURVE_INTENSITY = 64;
+
+export const MOUSE_BUTTON_LEFT = 0;
